% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FGCM}
\alias{FGCM}
\title{Fit logistic growth curve model}
\usage{
FGCM(
  geno,
  n.min,
  n.max,
  by = NULL,
  n_iter = NULL,
  multi.threads = TRUE,
  plot = TRUE
)
}
\arguments{
\item{geno}{Genotype information saved as a dataframe. Columns represent variants (SNPs or PCs).}

\item{n.min}{Minimum training set size}

\item{n.max}{Maximum training set size}

\item{by}{Default: NULL. Training set size tested from n.min and increases by this command. Divided evenly to 10 portions by default.}

\item{n_iter}{Number of simulation of each training set size. Automatically gave a suitable number by default.}

\item{multi.threads}{Default: TRUE. Set as FALSE if you just want to run it by single thread.}

\item{plot}{Default: TRUE. By default, this function will return a plot showing the simulation results and the fitted model.}
}
\value{
Estimation of parameters.
}
\description{
A function for fitting logisti growth model
}
