% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nt2r}
\alias{nt2r}
\title{Simulate r-scores of each training set size}
\usage{
nt2r(geno, peno, nt, n_iter = 30, n_core = NULL)
}
\arguments{
\item{geno}{A numeric dataframe of genotype, column represent sites (genotype coding as 1, 0, -1)}

\item{peno}{A numeric vector of phenotype}

\item{nt}{Numeric. Number of training set size.}

\item{n_iter}{Times of iteration. (default = 30)}

\item{n_core}{Number of cores. (Only tested on MacOS and UNIX-like system)}
}
\value{
A vector of r-scores of each iteration
}
\description{
Calculate r-scores (un-target) by in parallel.
}
