% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RErs.det}
\alias{RErs.det}
\title{Sample size determination for genomic prediction}
\usage{
RErs.det(geno, nt = NULL, n_iter = NULL, multi.threads = TRUE)
}
\arguments{
\item{geno}{A numeric data frame carried genotype information (column: variants/PCs, row: sample)}

\item{nt}{A numeric vector carried training set sizes for r-score simulation.}

\item{n_iter}{Number of iterations for estimating parameters.}

\item{multi.threads}{Default (multi.threads = TRUE) use 75\% of threads if the computer has more than 4 threads.}
}
\value{
An operating curve and its information.
}
\description{
Generate an operating curve for determining sample size.
}
