% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RErs.det}
\alias{RErs.det}
\title{Training set size determination}
\usage{
RErs.det(geno, nt = NULL, n_iter = NULL, multi.threads = TRUE)
}
\arguments{
\item{geno}{Genetic data performed by data.frame format}

\item{nt}{Vector of training set sizes for estimate logistic growth curve parameters. Auto generated by default = NULL.}

\item{n_iter}{Number of iterations for estimating parameters. Default times of iteration equals to the length of vector nt.}

\item{multi.threads}{Default (multi.threads = TRUE) use 75\% of threads if the computer has more than 4 threads.}
}
\value{
RErs result
}
\description{
Training set size determination for genomic selection. Detail described in published article by Wu et al.
}
